+++ src/main/java/view/element/CreatePartDialog.java
@@ -21,6 +21,7 @@ import java.awt.Insets;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.BorderFactory;
+import javax.swing.BoxLayout;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
 import javax.swing.JDialog;
@@ -54,6 +55,10 @@ public class CreatePartDialog extends JDialog implements ActionListener{
     
     protected StockItem currentStockItem;
     
+    protected JCheckBox assignToPart;
+    protected JCheckBox assignToSupplyer;
+    protected JPanel assignBC;
+    
     protected CreatePartController controller;
     public CreatePartDialog(CreatePartController cont, Frame owner, String titre, boolean modal){
         super(owner, titre, modal);
@@ -80,7 +85,10 @@ public class CreatePartDialog extends JDialog implements ActionListener{
         subPartCB.setSelectedIndex(0);
         minNum.setValue(0);
         supplyerCB.setSelectedIndex(0);
-        manufacturerCB.setSelectedIndex(0);     
+        manufacturerCB.setSelectedIndex(0);   
+        
+        assignToPart.setSelected(false);
+        assignToSupplyer.setSelected(true);
         
     }
     private void startSaveProcess(){
@@ -89,22 +97,26 @@ public class CreatePartDialog extends JDialog implements ActionListener{
         if(!UTILS.checkTextField(descTxt))
             status = false;
         
-        if(!UTILS.checkTextField(manufMPN))
-            status = false;
-        if(!UTILS.checkComboBox(manufacturerCB))
-                status = false;
-        /*
-        if(!UTILS.checkTextField(supplyerSKU))
-            status = false;
-        if(!UTILS.checkComboBox(supplyerCB)){
-            if(!UTILS.checkComboBox(manufacturerCB)){
-                status = false;
-            }else{
-                 supplyerCB.setBorder(BorderFactory.createLineBorder(Color.LIGHT_GRAY));
-            }
-         }else{
-            manufacturerCB.setBorder(BorderFactory.createLineBorder(Color.LIGHT_GRAY));
-        }*/
+        /*if(!UTILS.checkTextField(manufMPN))
+            status = false;*/
+        
+        /*if(!UTILS.checkComboBox(manufacturerCB))
+                status = false;*/
+       
+        if(!assignToPart.isSelected() && !assignToSupplyer.isSelected()){
+            UTILS.setCheckBorder(assignBC, false);
+             status = false;
+        }else{
+            UTILS.setCheckBorder(assignBC, true);
+        }
+        
+        if(assignToSupplyer.isSelected()){
+           System.out.println("YEAHH ");
+           Boolean test = UTILS.checkComboBox(supplyerCB);
+           UTILS.setCheckBorder(assignBC,  test);
+           status = status && test; 
+       }
+        
         
         if(!status){
             JOptionPane.showMessageDialog(this, "Some Required Information are missing");
@@ -124,7 +136,9 @@ public class CreatePartDialog extends JDialog implements ActionListener{
                 ((InventreeItem) supplyerCB.getSelectedItem()).getId(),
                 supplyerSKU.getText(),
                 ((InventreeItem) manufacturerCB.getSelectedItem()).getId(),
-                manufMPN.getText()
+                manufMPN.getText(),
+                assignToPart.isSelected(),
+                assignToSupplyer.isSelected()
         );
         
            
@@ -178,6 +192,13 @@ public class CreatePartDialog extends JDialog implements ActionListener{
         manufacturerCB = new InventreeItemFilterCB();
         manufMPN = new JTextField(txtFSize);
         saveBtn = new JButton("Save");
+        assignToPart=new JCheckBox("Assign Barcode to Part");
+        assignToSupplyer=new JCheckBox("Assign Barcode to Supplyer");
+        // pour panneau assignement
+        assignBC = new JPanel();
+        assignBC.setLayout(new BoxLayout(assignBC, BoxLayout.PAGE_AXIS));
+        assignBC.add(assignToPart);
+        assignBC.add(assignToSupplyer);
         
         //min size txtfield
         Dimension txtDim = new Dimension(200,20);
@@ -193,11 +214,14 @@ public class CreatePartDialog extends JDialog implements ActionListener{
         supplyerSKU.setMinimumSize(txtDim);
         manufacturerCB.setMinimumSize(txtDim);
         manufMPN.setMinimumSize(txtDim);
+        txtDim = new Dimension(200,50);
+        assignBC.setMinimumSize(txtDim);
         // panel
         JPanel jp = new JPanel(new GridBagLayout());
         GridBagConstraints cst = new GridBagConstraints();
         cst.anchor = GridBagConstraints.EAST;
         cst.insets = new Insets(0,0,10,10);
+        
         // les label
         int i = 0;
         cst.gridx = 0;
@@ -224,9 +248,12 @@ public class CreatePartDialog extends JDialog implements ActionListener{
         cst.gridy = (i+=2);
         jp.add(new JLabel("Manufacturer"), cst);
         cst.gridy = (i+=2);
-        jp.add(new JLabel("Manufacturer Id (set as barcode)"), cst);
+        jp.add(new JLabel("Manufacturer Id"), cst);
+        cst.gridy = (i+=2);
+        jp.add(new JLabel("Barcode Assignement"), cst);
         
         cst.anchor = GridBagConstraints.WEST;
+        cst.fill = GridBagConstraints.HORIZONTAL;
         i=0;
         cst.gridx=1;
         cst.gridy=i;
@@ -253,8 +280,11 @@ public class CreatePartDialog extends JDialog implements ActionListener{
         jp.add(manufacturerCB, cst);
         cst.gridy = (i+=2);
         jp.add(manufMPN, cst);
+        cst.gridy = (i+=2);
+        jp.add(assignBC, cst);        
         
         cst.anchor=GridBagConstraints.EAST;
+        cst.fill = GridBagConstraints.NONE;
         cst.gridy=(i+=1);
         jp.add(saveBtn, cst);