+++ changelog.md
@@ -1,3 +1,9 @@
+# 16-09-2022 / Beta v.0.6 :
+  * Fit with barcode api refactoring from Inventree 
+      * add barcode link to part, supplyer part
+      * remove barcode storage in MPN
+  * update readme
+
 # 12-09-2022 / Beta v.0.5 :
   * Add Changelog!
   * update readme